Refactor framegraph pass entry points for backend dispatch
