+++ Quake/gl_rmain.c
@@ -5668,252 +5668,6 @@ void R_WarpScaleView (const RenderGraphResourceHandle *resources)
 		framesetup.composite_ready = true;
 }
 
-static void R_FG_ExecSetupView (RenderPassContext *ctx)
-{
-	(void)ctx;
-	R_SetupView ();
-}
-
-static const RenderPassDesc s_setup_view_framegraph_pass = {
-	"Setup view",
-	RENDER_RES_NONE,
-	RENDER_RES_NONE,
-	1u << 0,
-	FG_PASS_OUTPUT_KEEP,
-	FG_PASS_VIEWPORT_KEEP,
-	NULL,
-	R_FG_ExecSetupView,
-	FG_PASS_STAGE_SETUP,
-	FG_PASS_STATS_SETUP
-};
-
-static qboolean R_FG_PassWhenShadowEnabled (const RenderPassContext *ctx)
-{
-	return ctx && ctx->frame_plan
-		&& ctx->frame_plan->run_shadowmaps;
-}
-
-static void R_FG_ExecShadowMaps (RenderPassContext *ctx)
-{
-	(void)ctx;
-	R_RenderShadowMaps ();
-}
-
-static const RenderPassDesc s_shadowmaps_framegraph_pass = {
-	"Shadow maps",
-	RENDER_RES_NONE,
-	RENDER_RES_SHADOW_SUN_DEPTH,
-	0,
-	FG_PASS_OUTPUT_KEEP,
-	FG_PASS_VIEWPORT_KEEP,
-	R_FG_PassWhenShadowEnabled,
-	R_FG_ExecShadowMaps,
-	FG_PASS_STAGE_MAIN,
-	FG_PASS_STATS_SHADOW
-};
-
-static void R_FG_ExecScene (RenderPassContext *ctx)
-{
-	R_RenderScene (ctx ? ctx->resources : NULL);
-}
-
-static const RenderPassDesc s_scene_framegraph_pass = {
-	"Render scene",
-	RENDER_RES_DECALS | RENDER_RES_SHADOW_SUN_DEPTH,
-	RENDER_RES_SCENE_COLOR | RENDER_RES_SCENE_DEPTH | RENDER_RES_VELOCITY,
-	0,
-	FG_PASS_OUTPUT_AUTO_SCENE,
-	FG_PASS_VIEWPORT_VIEW_RECT_SCALED,
-	NULL,
-	R_FG_ExecScene,
-	FG_PASS_STAGE_MAIN,
-	FG_PASS_STATS_SCENE
-};
-
-static void R_FG_ExecWarpResolve (RenderPassContext *ctx)
-{
-	R_WarpScaleView (ctx ? ctx->resources : NULL);
-}
-
-static const RenderPassDesc s_warp_resolve_framegraph_pass = {
-	"Warp/resolve",
-	RENDER_RES_SCENE_COLOR | RENDER_RES_SCENE_DEPTH,
-	RENDER_RES_COMPOSITE_COLOR | RENDER_RES_COMPOSITE_DEPTH,
-	1u << 0,
-	FG_PASS_OUTPUT_AUTO_WARP,
-	FG_PASS_VIEWPORT_VIEW_RECT,
-	NULL,
-	R_FG_ExecWarpResolve,
-	FG_PASS_STAGE_MAIN,
-	FG_PASS_STATS_WARP
-};
-
-static qboolean R_FG_HasResources (const RenderPassContext *ctx, unsigned required)
-{
-	unsigned mask;
-	const RenderGraphResourceHandle *resources = ctx ? ctx->resources : NULL;
-
-	if (!resources)
-		return false;
-
-	for (mask = 1u; mask != 0; mask <<= 1)
-	{
-		render_backend_resource_slot_t slot = R_BACKEND_RESOURCE_SLOT_NONE;
-		if ((required & mask) == 0)
-			continue;
-
-		switch (mask)
-		{
-		case RENDER_RES_SCENE_COLOR: slot = R_BACKEND_RESOURCE_SLOT_SCENE_COLOR; break;
-		case RENDER_RES_SCENE_DEPTH: slot = R_BACKEND_RESOURCE_SLOT_SCENE_DEPTH; break;
-		case RENDER_RES_COMPOSITE_COLOR: slot = R_BACKEND_RESOURCE_SLOT_COMPOSITE_COLOR; break;
-		case RENDER_RES_COMPOSITE_DEPTH: slot = R_BACKEND_RESOURCE_SLOT_COMPOSITE_DEPTH; break;
-		case RENDER_RES_SHADOW_SUN_DEPTH: slot = R_BACKEND_RESOURCE_SLOT_SHADOW_SUN_DEPTH; break;
-		case RENDER_RES_VELOCITY: slot = R_BACKEND_RESOURCE_SLOT_VELOCITY; break;
-		case RENDER_RES_DECALS:
-		case RENDER_RES_SSAO_FOG_STATE:
-			/* Internal logical framegraph resources; no backend slot backing. */
-			continue;
-		default:
-			return false;
-		}
-
-		if (!R_FrameGraph_HasResourceBySlot (resources, slot))
-			return false;
-	}
-
-	return true;
-}
-
-static void R_FG_ExecSSAOFogHandoff (RenderPassContext *ctx)
-{
-	(void)ctx;
-	/* Capture global fog parameters for deterministic postprocess SSAO suppression. */
-	R_SSAO_CaptureFogState (&r_framedata, &r_ssao_fog_state);
-}
-
-static const RenderPassDesc s_ssao_fog_handoff_framegraph_pass = {
-	"Capture fog handoff",
-	/* Captures CPU-side global fog into SSAO handoff data. */
-	RENDER_RES_NONE,
-	RENDER_RES_SSAO_FOG_STATE,
-	0,
-	FG_PASS_OUTPUT_KEEP,
-	FG_PASS_VIEWPORT_KEEP,
-	NULL,
-	R_FG_ExecSSAOFogHandoff
-};
-
-static qboolean R_FG_PassWhenPostprocessEnabled (const RenderPassContext *ctx)
-{
-	return ctx && ctx->frame_plan
-		&& ctx->frame_plan->run_postprocess
-		&& R_FG_HasResources (ctx, RENDER_RES_COMPOSITE_COLOR | RENDER_RES_COMPOSITE_DEPTH);
-}
-
-static void R_FG_ExecPostprocess (RenderPassContext *ctx)
-{
-	GL_PostProcess (ctx ? ctx->resources : NULL);
-}
-
-static const RenderPassDesc s_postprocess_framegraph_pass = {
-	"Postprocess",
-	RENDER_RES_COMPOSITE_COLOR | RENDER_RES_COMPOSITE_DEPTH | RENDER_RES_SSAO_FOG_STATE,
-	RENDER_RES_COMPOSITE_COLOR,
-	1u << 0,
-	FG_PASS_OUTPUT_BACKBUFFER,
-	FG_PASS_VIEWPORT_FULL_WINDOW,
-	R_FG_PassWhenPostprocessEnabled,
-	R_FG_ExecPostprocess,
-	FG_PASS_STAGE_MAIN,
-	FG_PASS_STATS_POST
-};
-
-static qboolean R_FG_PassWhenViewmodelEnabled (const RenderPassContext *ctx)
-{
-	return ctx && ctx->frame_plan && ctx->frame_plan->run_viewmodel;
-}
-
-static void R_FG_ExecViewmodel (RenderPassContext *ctx)
-{
-	(void)ctx;
-	R_DrawViewModel ();
-}
-
-static const RenderPassDesc s_viewmodel_framegraph_pass = {
-	"Draw viewmodel",
-	RENDER_RES_COMPOSITE_COLOR,
-	RENDER_RES_COMPOSITE_COLOR,
-	1u << 0,
-	FG_PASS_OUTPUT_BACKBUFFER,
-	FG_PASS_VIEWPORT_VIEW_RECT,
-	R_FG_PassWhenViewmodelEnabled,
-	R_FG_ExecViewmodel,
-	FG_PASS_STAGE_MAIN,
-	FG_PASS_STATS_OVERLAY
-};
-
-static qboolean R_FG_PassWhenPolyblendEnabled (const RenderPassContext *ctx)
-{
-	return ctx && ctx->frame_plan && ctx->frame_plan->run_polyblend;
-}
-
-static void R_FG_ExecPolyblend (RenderPassContext *ctx)
-{
-	(void)ctx;
-	V_PolyBlend ();
-}
-
-static const RenderPassDesc s_polyblend_framegraph_pass = {
-	"Polyblend",
-	RENDER_RES_COMPOSITE_COLOR,
-	RENDER_RES_COMPOSITE_COLOR,
-	1u << 0,
-	FG_PASS_OUTPUT_BACKBUFFER,
-	FG_PASS_VIEWPORT_VIEW_RECT,
-	R_FG_PassWhenPolyblendEnabled,
-	R_FG_ExecPolyblend,
-	FG_PASS_STAGE_MAIN,
-	FG_PASS_STATS_OVERLAY
-};
-
-static qboolean R_FG_PassWhenStorePrevEnabled (const RenderPassContext *ctx)
-{
-	return ctx && ctx->frame_plan && ctx->frame_plan->run_store_prev;
-}
-
-static void R_FG_ExecStorePrev (RenderPassContext *ctx)
-{
-	(void)ctx;
-	r_frame_rendered_this_update = true;
-	R_StorePrevFrameState ();
-}
-
-static const RenderPassDesc s_storeprev_framegraph_pass = {
-	"Store previous frame",
-	RENDER_RES_COMPOSITE_COLOR | RENDER_RES_SCENE_DEPTH,
-	RENDER_RES_NONE,
-	1u << 0,
-	FG_PASS_OUTPUT_KEEP,
-	FG_PASS_VIEWPORT_KEEP,
-	R_FG_PassWhenStorePrevEnabled,
-	R_FG_ExecStorePrev
-};
-
-void R_RegisterFrameGraphPasses (void)
-{
-	(void)R_FrameGraph_AddPass (&s_setup_view_framegraph_pass);
-	(void)R_FrameGraph_AddPass (&s_shadowmaps_framegraph_pass);
-	(void)R_FrameGraph_AddPass (&s_scene_framegraph_pass);
-	(void)R_FrameGraph_AddPass (&s_warp_resolve_framegraph_pass);
-	R_Decals_RegisterFrameGraphPasses ();
-	(void)R_FrameGraph_AddPass (&s_ssao_fog_handoff_framegraph_pass);
-	(void)R_FrameGraph_AddPass (&s_postprocess_framegraph_pass);
-	(void)R_FrameGraph_AddPass (&s_viewmodel_framegraph_pass);
-	(void)R_FrameGraph_AddPass (&s_polyblend_framegraph_pass);
-	(void)R_FrameGraph_AddPass (&s_storeprev_framegraph_pass);
-}
-
 /*
 ================
 R_RenderView
