+++ Windows/VisualStudio/ironwail.vcxproj
@@ -328,6 +328,7 @@ copy "$(SolutionDir)..\zlib\$(PlatformShortName)\*.dll" "$(TargetDir)"</Command>
     <ClCompile Include="..\..\Quake\r_dlight_pool.c" />
     <ClCompile Include="..\..\Quake\r_backend.c" />
     <ClCompile Include="..\..\Quake\r_framegraph.c" />
+    <ClCompile Include="..\..\Quake\r_passes.c" />
     <ClCompile Include="..\..\Quake\r_quality.c" />
     <ClCompile Include="..\..\Quake\r_sanitize.c" />
     <ClCompile Include="..\..\Quake\r_postfx.c" />
